fix: Test -> TradeList Container 명칭 변경, 삭제 된 글 표기 수정
