+++ src/app/pokemon-card-trade/page.js
@@ -1,11 +1,11 @@
-import TrradeContainer from "./components/TradeListContainer"
-import TestContainer from "./components/TestContainer"
+// import TempContainer from "./components/TempContainer"
+import TradeListContainer from "./components/TradeListContainer"
 
 export default function TradePage() {
   return (
     <>
       <div id="TradeList" className="w-[100%] flex flex-col gap-6">
-        <TestContainer />
+        <TradeListContainer />
       </div>
     </>
   );
